refactor(tasks): move G1 box-tracking owner [skip ci]
