+++ src/unilab/envs/motion_tracking/g1/__init__.py
@@ -1,12 +1,5 @@
 """Motion tracking environments for Unitree G1."""
 
-from .box_tracking import (
-    G1BoxTracking23DofCfg,
-    G1BoxTracking23DofEnvCfg,
-    G1BoxTrackingCfg,
-    G1BoxTrackingEnv,
-    G1BoxTrackingEnvCfg,
-)
 from .flip_tracking import (
     G1ClimbTrackingCfg,
     G1ClimbTrackingEnv,
@@ -32,7 +25,6 @@ from .flip_tracking_sac import (
     G1WallFlipTrackingSACCfg,
     G1WallFlipTrackingSACEnv,
 )
-from .motion_box_loader import BoxMotionData, BoxMotionLoader
 from .tracking import (
     G1MotionTracking23DofCfg,
     G1MotionTracking23DofDeployEnvCfg,
@@ -88,11 +80,4 @@ __all__ = [
     "G1ClimbTrackingCfg",
     "G1ClimbTrackingEnv",
     "G1ClimbTrackingEnvCfg",
-    "G1BoxTracking23DofCfg",
-    "G1BoxTracking23DofEnvCfg",
-    "G1BoxTrackingCfg",
-    "G1BoxTrackingEnv",
-    "G1BoxTrackingEnvCfg",
-    "BoxMotionData",
-    "BoxMotionLoader",
 ]
