+++ src/unilab/tasks/__init__.py
@@ -18,6 +18,7 @@ __unilab_registry_modules__ = (
     "unilab.tasks.manipulation.sharpa_inhand",
     "unilab.tasks.manipulation.stewart",
     "unilab.envs.motion_tracking.g1",
+    "unilab.tasks.motion_tracking.g1.box_tracking",
     "unilab.tasks.motion_tracking.x2",
 )
