+++ src/unilab/tasks/motion_tracking/g1/__init__.py
@@ -1,0 +1,20 @@
+"""G1 box-tracking tasks."""
+
+from .box_tracking import (
+    G1BoxTracking23DofCfg,
+    G1BoxTracking23DofEnvCfg,
+    G1BoxTrackingCfg,
+    G1BoxTrackingEnv,
+    G1BoxTrackingEnvCfg,
+)
+from .motion_box_loader import BoxMotionData, BoxMotionLoader
+
+__all__ = [
+    "BoxMotionData",
+    "BoxMotionLoader",
+    "G1BoxTracking23DofCfg",
+    "G1BoxTracking23DofEnvCfg",
+    "G1BoxTrackingCfg",
+    "G1BoxTrackingEnv",
+    "G1BoxTrackingEnvCfg",
+]
